feat: Refactor main app structure and integrate signup functionality with routing
